requirements.txt
openai==1.78.1
python-dotenv==1.0.0
